Stop using template haskell for config finder
The local config finder was generating the imports and a module list
and then we were using template haskell to actually create the list
at compile time. Now we just directly generate the correct haskell
source and include the generated file with the preprocessor.

The template haskell approach was not working well when the config
files were changed and recompiled. The compiled names would
sometimes change causing a link error in the previously generated
template haskell code.