Fix error handling for multiple errors
We were incorrectly reporting a successful run when there were
multiple checks to make for an output and one of the succeeded. When
there was no error we return Nothing. All the errors were gathered
in a list and 'sequence'ed together. A single Nothing would make it
appear as though no errors occured even if there were some Just
values in the list.