Add flag to conditionally build fibon-analyse
To support this we bumped Cabal dependency to 1.8. Versions before
1.8 did not properly seperate out the dependencies in the executable
sections. This means that all the build-depends were shared among
all the executables. See http://hackage.haskell.org/trac/hackage/ticket/742

With the later version of cabal we can conditionally build the
analyse program by wrapping its build-depends in an if condition.