Use a ByteString for the ExtraStats in FibonResult
This patch changes the Fibon.Result to use a ByteString for the
ExtraStats collected with each run. Earlier Result structures are
probably incompatible, but the show of a bytestring is the same as
the show of a string, so it is probably ok.

In addition, the fibon-analyse program was changed to use a custom
bytestring parser for constructing the GhcStats from the assoc list
output by the --machine-readable GHC stats.