Add a fibon->nofib build system converter
This commit adds the fibon-nofib tool. This tool can be used to ease
the transition from the fibon to the nofib build system. It attempts
to create a build environment that will work with the nofib build
system. It performs the following tasks:

1. Add a Makefile in the top level directory of each benchmark
2. Parse the output of cabal build (i.e. ghc --make) and add the
   required source file to the SRCS variable in the Makefile.
3. Read the Instance.hs and add appropiate entries to the HC_OPTS
   and PROG_ARGS Makefile variables.
4. Read the .cabal file and add the necessary -i and -package
   entries to the HC_OPTS variable.
5. Copy the input files and expected stdout files to the top level
   of each benchmark directory (from Fibon/data/...)

The generated Makefile may not be perfect (it does not include
necessary extensions), but it should be a good starting point for
hooking up to nofib.