Record configuration details in the log file
Record the final configuration details in the log file after
processing the users config file. The details are recorded in the
log file and as a separate .CONFIG file.

This commit also contains a refactoring of the log file
initialization. The timestamp and logfile path printing has been
moved into the Log module.