Add support for running executable with a script
This commit adds support for a `useRunScript` option in the config
file. The run script can be specified with its own args and then
will be called with its own args plus the executable and any run
flags it specifies.

Having an option for a run script is useful for running the
benchmarks with tools like dynamorio or pin that need to take
control of the executable.