Add stg timer stats to fibon-analyze
This commit makes the ghc stats parser aware of the experimental stg
timer. The stg timer counts the time that a ghc program spends
executing code for the stg machine. We add support for stg timer
output in the ghc stats file along with a basic table to display the
results.