Added Train as a new input size
The Train input size can be used for feedback based optimizations.
We plan to increase the standard running time for the Ref size so
we can work with longer running programs. The train input size is
therefore an option for when we want a quicker run but still get
some meaningful results.

Also updated fibon-init to include new Train data size.