Added a helper module for writing new benchmarks
The BenchmarkHelper module contains useful functions for writing
long running benchmarks. It contains a utility for parsing command
line arguments that strips out a -r N argument that can be used
internally by a program to run multiple times. Also, it exposes
the deepseq function so that benchmarks can use it to evaluate
data without printing it out which is useful when the program will
loop many times.