Add ability to reuse build results from old runs
We can now reuse the results of builds from old runs. Using
the old build results allow us to build the program under a
given configuration one time and then run it and collect
results many times. This feature is useful if you want to
collect performance numbers for configuration paramaters
that are passed at runtime instead of build time (for
example running with a different number of threads).

To reuse the build results pass the directory name as it
appers under the run/ directory in the fibon root. For
example,

  $ fibon-run -c head -r 401.head

will result the build results from run/401.head