Spawn cabal processes with correct working dir
This commit changes the way that cabal processes are spawned
so that they start with the correct working directory.
Previously we first changed the working directory of the
parent process and then spawned the cabal process. Doing it
that way prevents us from spawning multiple cabal processes
in parallel on different threads since the working directory
is a shared resource among multiple threads. We should now
be able to run multiple cabal processes which would let us
parallelize the build.