Run build steps in parallel
We can now run the build steps in parallel to speed the
overall collection of results. To support parallel builds we
now depend on the parallel-io package that provides a
thread pool for executing IO actions. To control the number
of threads used, you must pass the rts flags such as:

  $ fibon-run +RTS -N2 -RTS

This commit also introduces a dependency on the unix
package. We had to use this to install our own signal
handler to get CTRL-C to be correctly handled so that it
immediatly shuts down the program. Without the custom signal
handler it only kills one thread and so it is very difficult
to have an early termination of the program.