Allow benchmarks to be run in parallel
Sometimes for testing purposes it is useful to run the
benchmarks in parallel as well as building them in parallel.
For collecting running times it does not make sense to do
this, but for testing that everything is working a parallel
run is ok.

We add a new -p flag that will let the benchmarks be run in
parallel if the the runtime has enough threads (e.g. we are
using +RTS -N).